Fix nudge operations incurring FP error from coordinate space conversions
Closes https://github.com/ppy/osu/issues/31915.

Reproduction of aforementioned issue requires 1280x720 resolution, which
should also be a good way to confirm that this does anything.

To me this is also equal-parts-bugfix, equal-parts-code-quality PR,
because tell me: what on earth was this code ever doing at
`ComposeBlueprintContainer` level? Nudging by one playfield-space-unit
doesn't even *make sense* in something like taiko or mania.
